+++ NEWS.md
@@ -2,6 +2,7 @@
 
 2024--05--26
 
+- Add option for centered-dot multiplication symbol, often used with comma decimal marker
 - Change default whitespace argument to render correctly in the README page on GitHub
 - Revise wrapr::stop_if_dot_args() in functions, for more useful error message
 - Expanded the number of examples